requirements.txt
# Core dependencies
streamlit>=1.31.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
google-cloud-storage>=2.13.0
google-cloud-vision>=3.5.0
google-generativeai>=0.3.0
tenacity==8.2.3
python-dotenv>=1.0.0
vertexai>=0.0.1
google-cloud-aiplatform>=1.38.0
Pillow>=10.0.0

# Image processing
opencv-python>=4.8.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Additional dependencies for image processing
matplotlib>=3.7.0
